Refactoring
- Move OperatorCatalogPlatform into operator_platform.go
- change the render code to standardize on
map[OperatorCatalogPlatform]CatalogMetadata so that we can capture
new information in this map as the code goes through
getImageDigests -> renderDCFunc -> plan. To do this, renderDCFunc now
uses this map as an in/out parameter.
- CatalogMetadata.catalogRef is now a single ref rather than a slice
